route_log 엔티티 추가

+++ src/main/java/iriro/saferoute/entity/RoutePointLogEntity.java
@@ -30,5 +30,5 @@ public class RoutePointLogEntity {
     private BigDecimal longitude;
 
     @Column(nullable = false)
-    private Integer sequence;
+    private Long sequence;
 }
